furthur reducing the memory footprint by only pre-allocating smaller initial data structure. The ealier larger allocation is tested to be not necessary.

+++ src/c/common.h
@@ -54,7 +54,7 @@
  #################################################################################$$
  */
 
-typedef long int seq_coor_t; 
+typedef int seq_coor_t; 
 
 typedef struct {    
     seq_coor_t aln_str_size ;
